Skip writing client config if it is not writable
In some cases, the config file may be read-only. Since it is not
imperative that it be written (everything continues to work, it just
will not remember this device ID after a restart), just skip writing.

Ideally static and dynamic configuration should be totally separated,
however that is more complicated to implement and this is a quick fix.